# apps/backend/docs/swagger.yaml
definitions:
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.UpdateUserRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - id
    - password
    type: object
  user.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  user.UsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/user.UserResponse'
        type: array
    type: object
info:
  contact: {}
paths:
  /api/users:
    get:
      description: ユーザーの一覧を返します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UsersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザー一覧取得
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 新しいユーザーを作成します
      parameters:
      - description: ユーザー作成リクエスト
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザー作成
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 既存のユーザー情報を更新します
      parameters:
      - description: ユーザー更新リクエスト
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザー更新
      tags:
      - users
  /api/users/{id}:
    delete:
      description: 指定したユーザーを削除します
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザー削除
      tags:
      - users
  /api/users/email/{email}:
    get:
      description: ユーザーEmailを指定してユーザー情報を取得します
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザーEmailで取得
      tags:
      - users
  /api/users/id/{id}:
    get:
      description: ユーザーIDを指定してユーザー情報を取得します
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザーIDで取得
      tags:
      - users
swagger: "2.0"